refactor(ensnode-sdk): depend on datasources package
