+++ packages/datasources/package.json
@@ -46,7 +46,6 @@
     "viem": "catalog:"
   },
   "dependencies": {
-    "@ensnode/ensnode-sdk": "workspace:*",
     "@ponder/utils": "catalog:"
   }
 }
